CP-36392: Cast operands of divisions to not lose precision
Cast to `decimal` or `float` to keep floating-point precision

+++ XenAdmin/Controls/XenSearch/QueryElement.cs
@@ -2498,7 +2498,7 @@ namespace XenAdmin.Controls.XenSearch
                 if (intQuery == null)
                     return;
 
-                queryElement.numericUpDown.Value = intQuery.query / multiplier;
+                queryElement.numericUpDown.Value = (decimal) intQuery.query / multiplier;
             }
         }
